+++ packages/substrate-client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Avoid runtime error when the reply to `transaction_v1_broadcast` comes synchronously.
+
 ## 0.4.3 - 2025-08-08
 
 ### Fixed